Improve reliability of selecting correct track profile for each track shape
